Found bug in pcg (could not handle trivial case) and fixed it. Added trivial case test for pcg in test_pcg (test5). Corrected small error in time integrator when returning number of netwon iterations. Cleaned up test exits.

+++ src/Mesh.cpp
@@ -10,7 +10,7 @@ void Mesh::assign_masses(double rho) {
     size_t n = this->initial_positions.rows();
     size_t d = this->initial_positions.cols();
 
-    double node_mass = (rho * this->nx * this->ny) / (n * d);   // total mass divided by num nodes
+    double node_mass = (rho * this->nx * this->ny) / static_cast<double>(n * d);   // total mass divided by num nodes
 
     // Create COO coordinate triplet vector
     std::vector<Eigen::Triplet<double>> triplets;
